+++ apps/web/core/hooks/pages/use-extended-editor-extensions.ts
