+++ apps/web/core/hooks/pages/use-pages-pane-extensions.ts
